@@ -4,19 +4,19 @@
     <parent>
         <groupId>com.solace.maas</groupId>
         <artifactId>maas-event-management-agent-parent</artifactId>
-        <version>1.6.26-SNAPSHOT</version>
+        <version>1.6.27-SNAPSHOT</version>
         <relativePath>..</relativePath>
     </parent>
     <artifactId>event-management-agent</artifactId>
-    <version>1.6.26-SNAPSHOT</version>
+    <version>1.6.27-SNAPSHOT</version>
     <packaging>jar</packaging>
     <name>Solace Event Management Agent - Application</name>
     <description>Solace Event Management Agent - Application</description>
     <properties>
-        <spring-boot.version>3.3.10</spring-boot.version>
+        <spring-boot.version>3.4.4</spring-boot.version>
         <snakeyaml.version>2.0</snakeyaml.version>
         <spring-security-rsa.version>1.1.3</spring-security-rsa.version>
-        <spring-kafka.version>3.2.6</spring-kafka.version>
+        <spring-kafka.version>3.3.4</spring-kafka.version>
         <kafka-clients.version>3.8.1</kafka-clients.version>
         <jackson.version>2.16.1</jackson.version>
         <awaitility.version>4.2.0</awaitility.version>
@@ -267,32 +267,32 @@
         <dependency>
             <groupId>com.solace.maas</groupId>
             <artifactId>plugin</artifactId>
-            <version>1.6.26-SNAPSHOT</version>
+            <version>1.6.27-SNAPSHOT</version>
         </dependency>
         <dependency>
             <groupId>com.solace.maas.plugin.kafka</groupId>
             <artifactId>kafka-plugin</artifactId>
-            <version>1.6.26-SNAPSHOT</version>
+            <version>1.6.27-SNAPSHOT</version>
         </dependency>
         <dependency>
             <groupId>com.solace.maas.plugin.solace</groupId>
             <artifactId>solace-plugin</artifactId>
-            <version>1.6.26-SNAPSHOT</version>
+            <version>1.6.27-SNAPSHOT</version>
         </dependency>
         <dependency>
             <groupId>com.solace.maas.plugin.localstorage</groupId>
             <artifactId>local-storage-plugin</artifactId>
-            <version>1.6.26-SNAPSHOT</version>
+            <version>1.6.27-SNAPSHOT</version>
         </dependency>
         <dependency>
             <groupId>com.solace.maas.plugin.confluent-schema-registry</groupId>
             <artifactId>confluent-schema-registry-plugin</artifactId>
-            <version>1.6.26-SNAPSHOT</version>
+            <version>1.6.27-SNAPSHOT</version>
         </dependency>
         <dependency>
             <groupId>com.solace.maas.plugin.terraform</groupId>
             <artifactId>terraform-plugin</artifactId>
-            <version>1.6.26-SNAPSHOT</version>
+            <version>1.6.27-SNAPSHOT</version>
         </dependency>
 
         <dependency>
