@@ -19,8 +19,8 @@ import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.mock.mockito.SpyBean;
 import org.springframework.test.context.ActiveProfiles;
+import org.springframework.test.context.bean.override.mockito.MockitoSpyBean;
 
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
@@ -60,7 +60,7 @@ public class SempDeleteCommandManagerTest {
     private final static String DIR_SEMP_RESOURCES = "src/test/resources/sempResponses/";
     private final static String SEMP_RESPONSE_MISSING_RESOURCE = "sempResponseMissingResource.json";
 
-    @SpyBean
+    @MockitoSpyBean
     private SempDeleteCommandManager sempDeleteCommandManager;
 
     private SempApiProvider sempApiProvider;
