@@ -16,7 +16,6 @@ import lombok.extern.slf4j.Slf4j;
 public class SempApiProviderImpl implements SempApiProvider {
 
     private final ApiClient apiClient;
-    private final EventPortalProperties eventPortalProperties;
     private AclProfileApi aclProfileApi;
     private AuthorizationGroupApi authorizationGroupApi;
     private ClientUsernameApi clientUsernameApi;
@@ -25,8 +24,7 @@ public class SempApiProviderImpl implements SempApiProvider {
 
     public SempApiProviderImpl(SolaceHttpSemp solaceClient,
                                EventPortalProperties eventPortalProperties) {
-        this.apiClient = setupApiClient(solaceClient);
-        this.eventPortalProperties = eventPortalProperties;
+        this.apiClient = setupApiClient(solaceClient, eventPortalProperties);
     }
 
     @Override
@@ -61,16 +59,13 @@ public class SempApiProviderImpl implements SempApiProvider {
         return queueApi;
     }
 
-    private ApiClient setupApiClient(SolaceHttpSemp solaceClient) {
+    private ApiClient setupApiClient(SolaceHttpSemp solaceClient, EventPortalProperties eventPortalProperties) {
         SempClient sempClient = solaceClient.getSempClient();
         ApiClient client = new ApiClient();
         client.setBasePath(sempClient.getConnectionUrl() + "/SEMP/v2/config");
         client.setUsername(sempClient.getUsername());
         client.setPassword(sempClient.getPassword());
-        boolean verifyTls = eventPortalProperties == null || !eventPortalProperties.getSkipTlsVerify();
-        log.debug("SetVerifyingSsl? {} (application properties skipTlsVerify: {})", verifyTls,
-                eventPortalProperties == null ? "false" : eventPortalProperties.getSkipTlsVerify());
-        client.setVerifyingSsl(verifyTls);
+        client.setVerifyingSsl(eventPortalProperties == null || !eventPortalProperties.getSkipTlsVerify());
         return client;
     }
 
